Moved search app server into module __init__ file
